Typed fields only for test. Removed FileMetadata unique flag
