+++ base/src/main/kotlin/com/smeup/dbnative/metadata/file/PropertiesSerializer.kt
@@ -19,10 +19,7 @@ package com.smeup.dbnative.metadata.file
 
 import com.smeup.dbnative.model.Field
 import com.smeup.dbnative.model.FileMetadata
-import com.smeup.dbnative.utils.TypedField
-import com.smeup.dbnative.utils.TypedMetadata
 import com.smeup.dbnative.utils.fieldsToProperties
-import com.smeup.dbnative.utils.getFieldTypeInstance
 import java.io.File
 import java.io.FileInputStream
 import java.io.InputStreamReader
@@ -34,13 +31,25 @@ import kotlin.collections.ArrayList
 object PropertiesSerializer {
 
     fun propertiesToMetadata(propertiesDirPath: String, fileName: String): FileMetadata{
-        val mp: MutableMap<String, String> = loadProperties(propertiesDirPath, fileName)
+        val propertiesFile = FileInputStream(File("$propertiesDirPath${File.separatorChar}${fileName.toUpperCase()}.properties"))
+        //val properties = Properties()
+        //properties.load(InputStreamReader(propertiesFile, Charset.forName("UTF-8")))
+
+        val mp: MutableMap<String, String> = LinkedHashMap()
+        object : Properties() {
+            @Synchronized
+            override fun put(key: Any, value: Any): Any? {
+                return mp.put(key as String, value as String)
+            }
+        }.load(InputStreamReader(propertiesFile, Charset.forName("UTF-8")))
+
         // Fields
         var flds = mp.filterKeys { it.startsWith("field.") }
         val fields: MutableList<Field> = ArrayList()
         flds.forEach { fld ->
             val name = fld.key.split(".")[1]
-            val description = fld.value
+            val fldAttributes = fld.value.split(",")
+            val description = fldAttributes[0].trim()
             fields.add(Field(name, description))
         }
 
@@ -53,67 +62,18 @@ object PropertiesSerializer {
             fieldsKeys.addAll((mp.get("filekeys")?.split(",")!!))
         }
 
-        // Unique
-        val unique = mp.get("unique")!!.toBoolean()
 
-        return FileMetadata(fileName, recordFormat, fields, fieldsKeys, unique)
+
+        return FileMetadata(fileName, recordFormat, fields, fieldsKeys)
     }
 
-    fun propertiesToTypedMetadata(propertiesDirPath: String, fileName: String): TypedMetadata {
-        val mp: MutableMap<String, String> = loadProperties(propertiesDirPath, fileName)
-        // Fields
-        var flds = mp.filterKeys { it.startsWith("field.") }
-        val fields: MutableList<TypedField> = ArrayList()
-        flds.forEach { fld ->
-            val name = fld.key.split(".")[1]
-            val fldAttributes = fld.value.split(",")
-            val description = fldAttributes[0].trim()
-            val length = fldAttributes[2].trim().toInt()
-            val decimal = fldAttributes[3].trim().toInt()
 
-            val datatype = fldAttributes[1].trim()
-            val fieldType = datatype.getFieldTypeInstance(length, decimal)
 
-            fields.add(TypedField(Field(name,  description), fieldType))
-        }
-
-        // FormatName
-        val recordFormat = mp.get("recordformat")!!
-
-        // FieldKeys
-        val fieldsKeys: MutableList<String> = ArrayList()
-        if(!(mp.get("filekeys")).isNullOrEmpty()){
-            fieldsKeys.addAll((mp.get("filekeys")?.split(",")!!))
-        }
-
-        // Unique
-        val unique = mp.get("unique")!!.toBoolean()
-
-        return TypedMetadata(fileName, recordFormat, fields, fieldsKeys, unique)
-    }
-
-    private fun loadProperties(propertiesDirPath: String, fileName: String): MutableMap<String, String>{
-        val propertiesFile = FileInputStream(File("$propertiesDirPath${File.separatorChar}${fileName.toUpperCase()}.properties"))
-        //val properties = Properties()
-        //properties.load(InputStreamReader(propertiesFile, Charset.forName("UTF-8")))
-
-        val mp: MutableMap<String, String> = LinkedHashMap()
-        object : Properties() {
-            @Synchronized
-            override fun put(key: Any, value: Any): Any? {
-                return mp.put(key as String, value as String)
-            }
-        }.load(InputStreamReader(propertiesFile, Charset.forName("UTF-8")))
-        return mp;
-    }
 
     fun metadataToProperties(propertiesDirPath: String, fileMetadata: FileMetadata, overwrite: Boolean){
         _metadataToProperties(propertiesDirPath, fileMetadata, fileMetadata.fieldsToProperties(), overwrite)
     }
 
-    fun typedMetadataToProperties(propertiesDirPath: String, tMetadata: TypedMetadata, overwrite: Boolean){
-        _metadataToProperties(propertiesDirPath, tMetadata.fileMetadata(), tMetadata.fieldsToProperties(), overwrite)
-    }
 
     private fun _metadataToProperties(propertiesDirPath: String,
                                       fileMetadata: FileMetadata,
@@ -124,10 +84,6 @@ object PropertiesSerializer {
 
         var keys = "${fileMetadata.fileKeys.joinToString(",")}"
         properties.add(Pair("filekeys", keys))
-
-        var unique = "false"
-        if (fileMetadata.unique) unique = "true"
-        properties.add(Pair("unique", unique))
 
         val propertiesFilePath = "${propertiesDirPath}${File.separatorChar}${fileMetadata.tableName.toUpperCase()}.properties"
