+++ jt400/pom.xml
@@ -48,6 +48,13 @@
             <artifactId>base</artifactId>
             <version>1.1.0-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>com.smeup.reload</groupId>
+            <artifactId>base</artifactId>
+            <version>1.1.0-SNAPSHOT</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
         <!-- https://mvnrepository.com/artifact/net.sf.jt400/jt400 -->
         <dependency>
             <groupId>net.sf.jt400</groupId>
