+++ jt400/src/test/kotlin/com/smeup/dbnative/jt400/JT400OperationsOnFile.kt
@@ -24,7 +24,7 @@ import com.smeup.dbnative.model.DecimalType
 import com.smeup.dbnative.jt400.utils.createAndPopulateMunicipalityTable
 import com.smeup.dbnative.jt400.utils.dbManagerForTest
 import com.smeup.dbnative.jt400.utils.destroyDatabase
-import com.smeup.dbnative.utils.getField
+import com.smeup.dbnative.utils.propertiesToTypedMetadata
 import org.junit.*
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
@@ -224,7 +224,7 @@ class JT400OperationsOnFile {
         // Step2: read above written records and update A§DEA2 field
         // Step3: check and check for correct update
         // Step4: delete written record.
-        val tMetadata = PropertiesSerializer.propertiesToTypedMetadata("src/test/resources/dds/properties/", "BRARTI1L")
+        val tMetadata = propertiesToTypedMetadata("src/test/resources/dds/properties/", "BRARTI1L")
         dbManager.registerMetadata(tMetadata.fileMetadata(), false)
         val dbFile = dbManager.openFile("BRARTI1L")
