+++ manager/pom.xml
@@ -50,6 +50,13 @@
         </dependency>
         <dependency>
             <groupId>com.smeup.reload</groupId>
+            <artifactId>base</artifactId>
+            <version>1.1.0-SNAPSHOT</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.smeup.reload</groupId>
             <artifactId>sql</artifactId>
             <version>1.1.0-SNAPSHOT</version>
             <scope>test</scope>
