+++ nosql/pom.xml
@@ -49,6 +49,13 @@
             <artifactId>base</artifactId>
             <version>1.1.0-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>com.smeup.reload</groupId>
+            <artifactId>base</artifactId>
+            <version>1.1.0-SNAPSHOT</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
         <!-- https://mvnrepository.com/artifact/org.mongodb/mongo-java-driver -->
         <dependency>
             <groupId>org.mongodb</groupId>
