+++ nosql/src/test/kotlin/com/smeup/dbnative/nosql/NoSQLMunicipalityTest.kt
@@ -393,7 +393,6 @@ class NoSQLMunicipalityTest {
         assertTrue(dbFile.setll(key3A))
         var count = 0
         while (dbFile.eof() == false) {
-            var readed = dbFile.readEqual(key3A)
             count++
         }
         assertEquals(32, count)
