+++ sql/pom.xml
@@ -49,6 +49,13 @@
             <version>1.1.0-SNAPSHOT</version>
         </dependency>
         <dependency>
+            <groupId>com.smeup.reload</groupId>
+            <artifactId>base</artifactId>
+            <version>1.1.0-SNAPSHOT</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
             <groupId>org.hsqldb</groupId>
             <artifactId>hsqldb</artifactId>
             <version>2.5.0</version>
